debug image shape

+++ src/vision_semantic_segmentation_node.py
@@ -83,7 +83,7 @@ class VisionSemanticSegmentationNode:
         image_out_resized = self.seg.segmentation(image_in_resized)
         # print(image_out_resized.shape, "-->", image_in.shape)
 
-        self.generate_and_publish_convex_hull(image_out_resized, msg.header.frame_id)
+        self.generate_and_publish_convex_hull(image_out_resized.astype(np.uint8), msg.header.frame_id)
 
         # NOTE: we use INTER_NEAREST because values are discrete labels
         image_out = cv2.resize(image_out_resized, (image_in.shape[1], image_in.shape[0]),
